pyproject.toml
[project]
name = "bigbangsim"
version = "0.1.0"
description = "Cinematic, scientifically accurate Big Bang simulation in real-time 3D"
requires-python = ">=3.11"
dependencies = [
    "moderngl==5.12.0",
    "moderngl-window==3.1.1",
    "PyGLM==2.8.3",
    "numpy>=2.2,<3",
    "scipy>=1.14,<2",
    "Pillow>=12",
]

[project.optional-dependencies]
dev = ["pytest>=9", "ruff>=0.15"]

[tool.pytest.ini_options]
testpaths = ["tests"]

[build-system]
requires = ["setuptools>=75"]
build-backend = "setuptools.build_meta"
